Relax conditions to retrieve LV debug info
Dx used to make assumptions about variable scopes. It assumed that
parameter scopes started at pc 0 of a method, and that after a "store"
local variables scope start at the next instruction.
These assumptions may not be true when the bytecode is modified, for
instance when instrumented with Jacoco.
This CL loosens the conditions that allow to retrieve local variable
debug info according to the pc.
This fixes problems with missing local variables and parameters when
debugging and prevents an infinite loop in LocalVariableExtractor.

I also incremented the dx version to 1.13

Bug: 32432143
Bug: 25998400
Test: manual
Change-Id: I3f37febe770854e77793c9a458bf381d182722de
Signed-off-by: mydongistiny <jaysonedson@gmail.com>